Implemented a dedicated component for the doctor list in FacilityHome
