+++ src/Components/Facility/FacilityHome.tsx
@@ -1,9 +1,8 @@
 import * as Notification from "../../Utils/Notifications.js";
 
 import AuthorizeFor, { NonReadOnlyUsers } from "../../Utils/AuthorizeFor";
-import { DoctorModal, FacilityModel } from "./models";
+import { FacilityModel } from "./models";
 import {
-  DOCTOR_SPECIALIZATION,
   FACILITY_FEATURE_TYPES,
   USER_TYPES,
   getBedTypes,
@@ -19,9 +18,7 @@ import ConfirmDialog from "../Common/ConfirmDialog";
 import ContactLink from "../Common/components/ContactLink";
 import CoverImageEditModal from "./CoverImageEditModal";
 import DialogModal from "../Common/Dialog";
-import { DoctorCapacity } from "./DoctorCapacity";
-import { DoctorIcon } from "../TeleIcu/Icons/DoctorIcon";
-import DoctorsCountCard from "./DoctorsCountCard";
+
 import Page from "../Common/components/Page";
 import RecordMeta from "../../CAREUI/display/RecordMeta";
 import Table from "../Common/components/Table";
@@ -35,6 +32,7 @@ import request from "../../Utils/request/request.js";
 import routes from "../../Redux/api.js";
 import useQuery from "../../Utils/request/useQuery.js";
 import { FacilityHomeTriage } from "./FacilityHomeTriage.js";
+import { FacilityDoctorList } from "./FacilityDoctorList.js";
 
 const Loading = lazy(() => import("../Common/Loading"));
 
@@ -54,9 +52,7 @@ export const FacilityHome = (props: any) => {
   const [openDeleteDialog, setOpenDeleteDialog] = useState(false);
   const [editCoverImage, setEditCoverImage] = useState(false);
   const [imageKey, setImageKey] = useState(Date.now());
-  const [totalDoctors, setTotalDoctors] = useState(0);
   const [bedCapacityModalOpen, setBedCapacityModalOpen] = useState(false);
-  const [doctorCapacityModalOpen, setDoctorCapacityModalOpen] = useState(false);
   const authUser = useAuthUser();
   const config = useConfig();
 
@@ -71,7 +67,6 @@ export const FacilityHome = (props: any) => {
       onResponse: ({ res }) => {
         if (res?.ok) {
           capacityQuery.refetch();
-          doctorQuery.refetch();
         } else {
           navigate("/not-found");
         }
@@ -81,21 +76,6 @@ export const FacilityHome = (props: any) => {
 
   const capacityQuery = useQuery(routes.getCapacity, {
     pathParams: { facilityId },
-  });
-
-  const doctorQuery = useQuery(routes.listDoctor, {
-    pathParams: { facilityId: facilityId },
-    onResponse: ({ res, data }) => {
-      if (res?.ok && data) {
-        let totalCount = 0;
-        [data].map((doctor: DoctorModal) => {
-          if (doctor.count) {
-            totalCount += doctor.count;
-          }
-        });
-        setTotalDoctors(totalCount);
-      }
-    },
   });
 
   const handleDeleteClose = () => {
@@ -177,68 +157,6 @@ export const FacilityHome = (props: any) => {
               />
             );
           }
-        })}
-      </div>
-    );
-  }
-
-  let doctorList: any = null;
-  if (!doctorQuery.data || !doctorQuery.data.results.length) {
-    doctorList = (
-      <h5 className="flex w-full items-center justify-center rounded-lg bg-white p-4 text-xl font-bold text-gray-500 shadow">
-        No Doctors Found
-      </h5>
-    );
-  } else {
-    doctorList = (
-      <div className="mt-4 grid gap-6 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
-        {/* Total Doctors Count Card */}
-        <div className="w-full">
-          <div className="flex h-full flex-col rounded-sm border border-primary-500 bg-primary-100 shadow-sm">
-            <div className="flex flex-1 items-center justify-start gap-3 px-4 py-6">
-              <div className="rounded-full bg-primary-500 p-4">
-                <DoctorIcon className="h-5 w-5 fill-current text-white" />
-              </div>
-              <div>
-                <div className="text-sm font-medium text-[#808080]">
-                  Total Doctors
-                </div>
-                <h2 className="mt-2 text-xl font-bold">{totalDoctors}</h2>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {doctorQuery.data.results.map((data: DoctorModal) => {
-          const removeCurrentDoctorData = (doctorId: number | undefined) => {
-            if (doctorQuery.data !== undefined) {
-              doctorQuery.data.results = doctorQuery.data?.results.filter(
-                (i: DoctorModal) => i.id !== doctorId
-              );
-            }
-          };
-
-          return (
-            <DoctorsCountCard
-              facilityId={facilityId}
-              key={`bed_${data.id}`}
-              handleUpdate={async () => {
-                doctorQuery.refetch();
-                if (doctorQuery.res?.ok && doctorQuery.data) {
-                  // update total doctors count
-                  let totalCount = 0;
-                  doctorQuery.data.results.map((doctor: DoctorModal) => {
-                    if (doctor.count) {
-                      totalCount += doctor.count;
-                    }
-                  });
-                  setTotalDoctors(totalCount);
-                }
-              }}
-              {...data}
-              removeDoctor={removeCurrentDoctorData}
-            />
-          );
         })}
       </div>
     );
@@ -629,21 +547,8 @@ export const FacilityHome = (props: any) => {
         </div>
         <div>{capacityList}</div>
       </div>
-      <div className="mt-5 rounded bg-white p-3 shadow-sm md:p-6">
-        <div className="justify-between md:flex md:pb-2">
-          <div className="mb-2 text-xl font-bold">Doctors List</div>
-          <ButtonV2
-            className="w-full md:w-auto"
-            onClick={() => setDoctorCapacityModalOpen(true)}
-            disabled={doctorList.length === DOCTOR_SPECIALIZATION.length}
-            authorizeFor={NonReadOnlyUsers}
-          >
-            <i className="fas fa-user-md mr-2 text-white" />
-            Add Doctor Types
-          </ButtonV2>
-        </div>
-        <div className="mt-4">{doctorList}</div>
-      </div>
+
+      <FacilityDoctorList facilityId={facilityId} />
 
       <div className="mt-5 rounded bg-white p-3 shadow-sm md:p-6">
         <FacilityHomeTriage
@@ -664,32 +569,6 @@ export const FacilityHome = (props: any) => {
             handleClose={() => setBedCapacityModalOpen(false)}
             handleUpdate={async () => {
               capacityQuery.refetch();
-            }}
-          />
-        </DialogModal>
-      )}
-      {doctorCapacityModalOpen && (
-        <DialogModal
-          show={doctorCapacityModalOpen}
-          onClose={() => setDoctorCapacityModalOpen(false)}
-          title="Add Doctor Capacity"
-          className="max-w-md md:min-w-[600px]"
-        >
-          <DoctorCapacity
-            facilityId={facilityId}
-            handleClose={() => setDoctorCapacityModalOpen(false)}
-            handleUpdate={async () => {
-              doctorQuery.refetch();
-              if (doctorQuery.res?.ok && doctorQuery.data) {
-                // update total doctors count
-                setTotalDoctors(
-                  doctorQuery.data.results.reduce(
-                    (acc: number, doctor: DoctorModal) =>
-                      acc + (doctor.count || 0),
-                    0
-                  )
-                );
-              }
             }}
           />
         </DialogModal>
